NIFI-4583 Restructure nifi-solr-processors
Make all methods static

NIFI-4583 Restructure nifi-solr-processors R2

This closes #2285.
